Split S3Common into multiple files
