+++ src/IO/S3/Client.h
@@ -10,7 +10,9 @@
 
 #include <IO/S3/URI.h>
 #include <IO/S3/Requests.h>
+#include <IO/S3/PocoHTTPClient.h>
 
+#include <aws/core/Aws.h>
 #include <aws/core/client/DefaultRetryStrategy.h>
 #include <aws/s3/S3Client.h>
 #include <aws/s3/S3ServiceClientModel.h>
@@ -300,6 +302,39 @@ private:
     const size_t max_redirects;
 
     Poco::Logger * log;
+};
+
+class ClientFactory
+{
+public:
+    ~ClientFactory();
+
+    static ClientFactory & instance();
+
+    std::unique_ptr<S3::Client> create(
+        const PocoHTTPClientConfiguration & cfg,
+        bool is_virtual_hosted_style,
+        const String & access_key_id,
+        const String & secret_access_key,
+        const String & server_side_encryption_customer_key_base64,
+        HTTPHeaderEntries headers,
+        bool use_environment_credentials,
+        bool use_insecure_imds_request);
+
+    PocoHTTPClientConfiguration createClientConfiguration(
+        const String & force_region,
+        const RemoteHostFilter & remote_host_filter,
+        unsigned int s3_max_redirects,
+        bool enable_s3_requests_logging,
+        bool for_disk_s3,
+        const ThrottlerPtr & get_request_throttler,
+        const ThrottlerPtr & put_request_throttler);
+
+private:
+    ClientFactory();
+
+    Aws::SDKOptions aws_options;
+    std::atomic<bool> s3_requests_logging_enabled;
 };
 
 }
